Fix truncation of values written by emulated code

+++ src/cpu/dynarmic_wrapper/lib.cpp
@@ -16,9 +16,9 @@ std::uint16_t touchHLE_cpu_read_u16(touchHLE_Mem *mem, VAddr addr, bool *error);
 std::uint32_t touchHLE_cpu_read_u32(touchHLE_Mem *mem, VAddr addr, bool *error);
 std::uint64_t touchHLE_cpu_read_u64(touchHLE_Mem *mem, VAddr addr, bool *error);
 bool touchHLE_cpu_write_u8(touchHLE_Mem *mem, VAddr addr, std::uint8_t value);
-bool touchHLE_cpu_write_u16(touchHLE_Mem *mem, VAddr addr, std::uint8_t value);
-bool touchHLE_cpu_write_u32(touchHLE_Mem *mem, VAddr addr, std::uint8_t value);
-bool touchHLE_cpu_write_u64(touchHLE_Mem *mem, VAddr addr, std::uint8_t value);
+bool touchHLE_cpu_write_u16(touchHLE_Mem *mem, VAddr addr, std::uint16_t value);
+bool touchHLE_cpu_write_u32(touchHLE_Mem *mem, VAddr addr, std::uint32_t value);
+bool touchHLE_cpu_write_u64(touchHLE_Mem *mem, VAddr addr, std::uint64_t value);
 }
 
 const auto HaltReasonSvc = Dynarmic::HaltReason::UserDefined1;
